chore: consolidate DB queries to single module
